test: add `test_proposer_defends_successfully` test
