+++ book/SUMMARY.md
@@ -28,6 +28,7 @@
   - [How to run the FP Proposer](./fault_proofs/proposer.md)
   - [How to run the FP Challenger](./fault_proofs/challenger.md)
   - [Gas Costs](./fault_proofs/gas_costs.md)
+  - [Testing Guide](./fault_proofs/testing.md)
 - [Experimental](./experimental/intro.md)
   - [OptimismPortalV2](./experimental/optimism-portal-v2.md)
 - [FAQ](./faq.md)
